refactor: consolidate gameweek manager interface and enhance functionality
- Removed the outdated gameweek manager interface file, streamlining the codebase for better maintainability.
- Updated the main gameweek manager to improve squad and transfer analysis features, including enhanced formatting for transfer deadline and data freshness displays.
- Improved error handling and user feedback mechanisms for loading gameweek data and displaying relevant information.
- Refined function signatures and consolidated imports to enhance clarity and reduce redundancy.
